+++ core/connectors/postgres/event_log_repo.js
@@ -85,3 +85,4 @@ class EventLogRepo {
 module.exports = { EventLogRepo };
 
 
+