+++ core/runtime/trace.js
@@ -8,3 +8,4 @@ function makeTraceId() {
 module.exports = { makeTraceId };
 
 
+