+++ core/tests/helpers_normalize.test.js
@@ -45,3 +45,4 @@ test('clampRating1to5: rounds and clamps', () => {
 });
 
 
+